sentinelpi/flask_ui/requirements.txt
Flask==2.1.2
Flask-MQTT==0.9.0
Flask-Cors==3.0.10
requests==2.26.0
paho-mqtt==1.6.1
python-telegram-bot==13.7
gunicorn==20.1.0
opencv-python-headless==4.5.3.20210927
tensorflow==2.6.0
torch==1.9.0+cpu
torchvision==0.10.0+cpu
numpy==1.21.2
Pillow==8.4.0
python-dotenv==0.19.2